Added DTMF callback support all the way to PJSUA API (ticket #48)
git-svn-id: https://svn.pjsip.org/repos/pjproject/trunk@863 74dad513-b988-da41-8d7b-12977e46ad98

+++ pjsip/src/pjsua-lib/pjsua_media.c
@@ -442,7 +442,7 @@ pj_status_t pjsua_media_subsys_destroy(void)
     }
 
     /* Close media transports */
-    for (i=0; i<(int)pjsua_var.ua_cfg.max_calls; ++i) {
+    for (i=0; i<pjsua_var.ua_cfg.max_calls; ++i) {
 	if (pjsua_var.calls[i].med_tp) {
 	    (*pjsua_var.calls[i].med_tp->op->destroy)(pjsua_var.calls[i].med_tp);
 	    pjsua_var.calls[i].med_tp = NULL;
